Fixed import conflicts between solvers (read description).
Some solvers such as Waves and Metafor contain modules of the same name, causing import conflicts (as all the solver paths were hard-coded in CUPyDO and added to sys.path).

The user must now indicate the solver path(s) in a json input file. The related path is automatically added to sys.path when importing a solver, then removed to prevent potential conflicts.