Fix PFEM3D "state" BC on Gaston (read description)
The previous implementation of the "State" BC (using the current state of the node as BC) is working on Windows but leads to incorrect results on Gaston.

In order to quickly fix the problem, the nodes are now storing their "State" BC explicitly in the class. In fine, one should come back to the previous implementation and investigate the problem on Gaston.